emails: Separate the utilities from the huey-powered backend
In order to avoid running connections to Huey in places that don't need
it.
